requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
httpx
rich
pytest
black
